Added proper threadpooling that generates and rebuilds chunks.
The chunk processing queue drops entries that are outdated so that the rebuilting doesn't waste resources.

can quickly generate 32*32 chunks on 3 threads in release mode

+++ MinecraftOpenGL/Source/Utils.cpp
@@ -5,7 +5,7 @@
 #include "World/Chunk/Chunk.h"
 
 int Utils::to1D(glm::vec3& position, int width) {
-	return position.x * width * width + position.y * width + position.z;
+	return (int)(position.x * width * width + position.y * width + position.z);
 }
 
 int Utils::to1D(glm::ivec2 position, int width) {
